% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hidden_aliases.R, R/MsBackendSqlDb.R
\docType{methods}
\name{hidden_aliases}
\alias{hidden_aliases}
\alias{$,MsBackendSqlDb-method}
\alias{$<-,MsBackendSqlDb-method}
\title{Internal page for hidden aliases}
\usage{
\S4method{$}{MsBackendSqlDb}(x, name)

\S4method{$}{MsBackendSqlDb}(x, name) <- value
}
\description{
For S4 methods that require a documentation entry but only clutter the index.
}
\keyword{internal}
