% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hidden_aliases.R, R/MsBackendSqlDb.R
\docType{methods}
\name{hidden_aliases}
\alias{hidden_aliases}
\alias{show,MsBackendSqlDb-method}
\alias{acquisitionNum,MsBackendSqlDb-method}
\alias{centroided,MsBackendSqlDb-method}
\alias{collisionEnergy,MsBackendSqlDb-method}
\alias{dataOrigin,MsBackendSqlDb-method}
\alias{dataStorage,MsBackendSqlDb-method}
\alias{intensity,MsBackendSqlDb-method}
\alias{ionCount,MsBackendSqlDb-method}
\alias{isCentroided,MsBackendSqlDb-method}
\alias{isEmpty,MsBackendSqlDb-method}
\alias{isolationWindowLowerMz,MsBackendSqlDb-method}
\alias{isolationWindowTargetMz,MsBackendSqlDb-method}
\alias{isolationWindowUpperMz,MsBackendSqlDb-method}
\alias{length,MsBackendSqlDb-method}
\alias{lengths,MsBackendSqlDb-method}
\alias{msLevel,MsBackendSqlDb-method}
\alias{mz,MsBackendSqlDb-method}
\alias{as.list,MsBackendSqlDb-method}
\alias{polarity,MsBackendSqlDb-method}
\alias{precScanNum,MsBackendSqlDb-method}
\alias{precursorCharge,MsBackendSqlDb-method}
\alias{precursorIntensity,MsBackendSqlDb-method}
\alias{precursorMz,MsBackendSqlDb-method}
\alias{rtime,MsBackendSqlDb-method}
\alias{scanIndex,MsBackendSqlDb-method}
\alias{smoothed,MsBackendSqlDb-method}
\alias{asDataFrame,MsBackendSqlDb-method}
\alias{spectraNames,MsBackendSqlDb-method}
\alias{spectraVariables,MsBackendSqlDb-method}
\alias{tic,MsBackendSqlDb-method}
\alias{$,MsBackendSqlDb-method}
\alias{$<-,MsBackendSqlDb-method}
\title{Internal page for hidden aliases}
\usage{
\S4method{show}{MsBackendSqlDb}(object)

\S4method{acquisitionNum}{MsBackendSqlDb}(object)

\S4method{centroided}{MsBackendSqlDb}(object)

\S4method{collisionEnergy}{MsBackendSqlDb}(object)

\S4method{dataOrigin}{MsBackendSqlDb}(object)

\S4method{dataStorage}{MsBackendSqlDb}(object)

\S4method{intensity}{MsBackendSqlDb}(object)

\S4method{ionCount}{MsBackendSqlDb}(object)

\S4method{isCentroided}{MsBackendSqlDb}(object, ...)

\S4method{isEmpty}{MsBackendSqlDb}(x)

\S4method{isolationWindowLowerMz}{MsBackendSqlDb}(object)

\S4method{isolationWindowTargetMz}{MsBackendSqlDb}(object)

\S4method{isolationWindowUpperMz}{MsBackendSqlDb}(object)

\S4method{length}{MsBackendSqlDb}(x)

\S4method{lengths}{MsBackendSqlDb}(x, use.names = FALSE)

\S4method{msLevel}{MsBackendSqlDb}(object, ...)

\S4method{mz}{MsBackendSqlDb}(object)

\S4method{as.list}{MsBackendSqlDb}(x)

\S4method{polarity}{MsBackendSqlDb}(object)

\S4method{precScanNum}{MsBackendSqlDb}(object)

\S4method{precursorCharge}{MsBackendSqlDb}(object)

\S4method{precursorIntensity}{MsBackendSqlDb}(object)

\S4method{precursorMz}{MsBackendSqlDb}(object)

\S4method{rtime}{MsBackendSqlDb}(object)

\S4method{scanIndex}{MsBackendSqlDb}(object)

\S4method{smoothed}{MsBackendSqlDb}(object)

\S4method{asDataFrame}{MsBackendSqlDb}(object, columns = spectraVariables(object))

\S4method{spectraNames}{MsBackendSqlDb}(object)

\S4method{spectraVariables}{MsBackendSqlDb}(object)

\S4method{tic}{MsBackendSqlDb}(object, initial = TRUE)

\S4method{$}{MsBackendSqlDb}(x, name)

\S4method{$}{MsBackendSqlDb}(x, name) <- value
}
\description{
For S4 methods that require a documentation entry but only clutter the index.
}
\keyword{internal}
